chore: Carousel Areas component added to fix duplicated lines
